+++ pages/manageFormula.php
@@ -546,14 +546,27 @@ if($_POST['action'] == 'makeFormula' && $_POST['fid'] && $_POST['q'] && $_POST['
 }
 //MARK COMPLETE
 if($_POST['action'] == 'todo' && $_POST['fid'] && $_POST['markComplete']){
+	require_once(__ROOT__.'/libs/fpdf.php');
+	require_once(__ROOT__.'/func/genBatchID.php');
+	require_once(__ROOT__.'/func/genBatchPDF.php');
+	require_once(__ROOT__.'/func/ml2L.php');
+
 	$fid = mysqli_real_escape_string($conn, $_POST['fid']);
+	$total_quantity = mysqli_real_escape_string($conn, $_POST['totalQuantity']);
+
+	define('FPDF_FONTPATH',__ROOT__.'/fonts');
+	$defCatClass = $settings['defCatClass'];
+	
+		
 	if(mysqli_num_rows(mysqli_query($conn, "SELECT id FROM makeFormula WHERE fid = '$fid' AND toAdd = '1'"))){
 		$response['error'] = '<strong>Formula is pending materials to add, cannot be marked as complete.</strong>';
 		echo json_encode($response);
 		return;
 	}
 	if(mysqli_query($conn,"UPDATE formulasMetaData SET isMade = '1', toDo = '0', madeOn = NOW(), status = '2' WHERE fid = '$fid'")){
-		
+		$batchID = genBatchID();
+		genBatchPDF($fid,$batchID,$total_quantity,'100',$total_quantity,$defCatClass,$settings['qStep'],'makeFormula');
+
 		mysqli_query($conn, "DELETE FROM makeFormula WHERE fid = '$fid'");
 		
 		$response['success'] = '<strong>Formula is complete</strong>';